Track Blood Moon and Incursions?
This is the best way I could come up with handling it. No real way to track it obviously but we know the BM starts ever 3 hours and ends in 30 mins.

We added a level req to them as well. Might need to be adjusted over time. If you are below said level (We have 10 level off the requirement for BM) then it wont show for you.